+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 All Notable changes to dbapi will be documented in this file
 
-Date: 0000-0-00
+Date: 2015-4-16
 
 ### Added
-Nothing
+- Table Creation